--- cluster_affinity/cluster_computation/__init__.py
@@ -1,1 +1,0 @@
-from .cluster_affinity import convert_tree_size_array
