+++ src/__init__.py
@@ -1,0 +1,2 @@
+from .cluster_computation import rooted_cluster_affinity
+from .main import cluster_affinity_script
